Fix undefined symbol warnings and build issues
- ecdsa.cpp: Add header include, rename internal Signature to SignatureData
  to avoid conflict with header, add C++ wrapper implementations for
  secp256k1Sign, secp256k1SignRecoverable, secp256k1Verify, p256Sign, p256Verify
- eddsa.cpp: Add header include, add ed25519Sign/ed25519Verify wrappers,
  fix C API function signatures to match header declarations
- secure_memory.cpp: Fix shift overflow warning by casting uintptr_t to uint64_t

+++ src/secure_memory.cpp
@@ -782,9 +782,9 @@ void MaskedKey<N>::initializeMask() {
         // Fallback: Use a deterministic but unpredictable pattern
         // This is less secure but better than nothing
         // Mix in the address of this object for some entropy
-        uintptr_t addr = reinterpret_cast<uintptr_t>(this);
+        uint64_t addr = static_cast<uint64_t>(reinterpret_cast<uintptr_t>(this));
         for (size_t i = 0; i < N; ++i) {
-            // Simple PRNG-like mixing
+            // Simple PRNG-like mixing (LCG)
             addr = addr * 6364136223846793005ULL + 1442695040888963407ULL;
             mask_[i] = static_cast<uint8_t>(addr >> 56);
         }
